Updated mbr tests to permissive range
